Common create_mnist_model
